Simplify dist folder for alpha
