+++ design-tokens/package.json
@@ -8,24 +8,18 @@
       "import": "./dist/esm/index.js",
       "require": "./dist/cjs/index.cjs"
     },
-    "./esm-unresolved": "./dist/esm-unresolved/index.js",
+    "./docs": "./dist/docs/index.js",
     "./dist/css/*.css": {
       "import": "./dist/css/*.css",
       "require": "./dist/css/*.css"
-    },
-    "./dist/tailwind/*.js": {
-      "import": "./dist/tailwind/*.js",
-      "require": "./dist/tailwind/*.js"
-    },
-    "./tree": "./dist/tree/index.js"
+    }
   },
   "scripts": {
     "prettier:check": "prettier --check src/",
     "sync-tokens-to-figma": "tsx src/scripts/sync_tokens_to_figma.ts",
     "sync-figma-to-tokens": "tsx src/scripts/sync_figma_to_tokens.ts",
     "build-style-dictionary": "tsx src/scripts/build-style-dictionary.js",
-    "build-token-tree": "tsx src/scripts/build-token-tree.js",
-    "build": "rm -rf ./dist && tsx src/scripts/build.js && yarn build-style-dictionary && yarn build-token-tree",
+    "build": "rm -rf ./dist && tsx src/scripts/build.js && yarn build-style-dictionary",
     "release-stable": "node src/scripts/release-stable.js  --branch=design-tokens-alpha-stable",
     "paddingY:verify": "tsx src/scripts/verify-paddingY.ts",
     "paddingY:update": "tsx src/scripts/update-paddingY.ts",
